+++ src/lib/connection/KnxLink/index.ts
@@ -5,6 +5,8 @@ export type {
     OnError,
     KnxDisconnectedReason,
     KnxConnectingEvent,
+    KnxNetworkConnectionEstablishedEvent,
+    KnxStartingSessionEvent,
     KnxReconnectingEvent,
     KnxDisconnectedEvent,
     LinkInfo